feat: shows failed commands in the terminal
BREAKING CHANGE:

- requires registering a task, and changes how things are tracked.
- requires Cypress v3.1.0+
