+++ cypress/support/index.js
@@ -1,2 +1,4 @@
 // load cypress-failed-log
+// in the user project this would be
+// require('cypress-failed-log')
 require('../..')
